Rename PlatformPlatform.SharedKernel.ApiCore.HttpResults to ApiResults
