+++ account-management/Api/Users/UserEndpoints.cs
@@ -1,7 +1,7 @@
 using MediatR;
 using PlatformPlatform.AccountManagement.Application.Users;
 using PlatformPlatform.AccountManagement.Domain.Users;
-using PlatformPlatform.SharedKernel.ApiCore.HttpResults;
+using PlatformPlatform.SharedKernel.ApiCore.ApiResults;
 
 namespace PlatformPlatform.AccountManagement.Api.Users;
